pyproject.toml
[project]
name = "aps-mcp-server-python"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=44.0.0",
    "fastmcp>=3.1.0",
    "httpx>=0.28.1",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared", "mcp_server_2lo", "mcp_server_ssa", "mcp_server_3lo"]
